make BearerTokenCredential an interface
